+++ Synapsis/branch/rev 82_fullStaticSynapsisCore/chainconst.h
@@ -1,8 +1,8 @@
 #ifndef CHAINCONST_H
 #define CHAINCONST_H
 const double PI=3.1415926535;
-const int maxa=910;
-const int maxcross=500;
+const long maxa=910;
+const long maxcross=500;
 
 //DNA torsional constant.
 //C=2.4~3e-19 erg.cm=2.4~3e-28 J.m. Change this unit to kT.basepairlength
